test(StackedBarChart): Fix tests for mapStackedBarchartData
