+++ src/components/StackedBarChart/index.tsx
@@ -2,8 +2,8 @@ import React, { FC, useEffect, useState, useMemo } from 'react';
 import styled from 'styled-components';
 import { SelectedTreeType } from '../../common/interfaces';
 import { Bars } from './Bars';
+import { getChartConfig } from './getChartConfig';
 import {
-  ChartConfigType,
   DayWaterAmountsType,
   mapStackedBarchartData,
 } from './mapStackedBarchartData';
@@ -28,29 +28,6 @@ const BarChartWrapper = styled.div`
   }
 `;
 
-const getChartConfig = (): ChartConfigType => {
-  const chartHeight = 140;
-  const chartWidth = 310;
-  const margins = {
-    top: 25,
-    right: 5,
-    bottom: 20,
-    left: 20,
-  };
-  const innerHeight = chartHeight - margins.top - margins.bottom;
-  const innerWidth = chartWidth - margins.left - margins.right;
-  return {
-    chartHeight,
-    chartWidth,
-    margins,
-    barPadding: 2,
-    innerHeight,
-    innerWidth,
-    barWidth: Math.round(innerWidth / 30),
-    minYValue: 60,
-    tickSize: 6,
-  };
-};
 const config = getChartConfig();
 
 const StackedBarChart: FC<{
